Notificaciones basicas exitosas y avanze de la pantalla configuracion

+++ app/src/main/java/com/example/neolink_app/clases/configuracion/Confvalues.java
@@ -3,20 +3,20 @@ package com.example.neolink_app.clases.configuracion;
 public class Confvalues {
     public double BAT_H;
     public double BAT_L;
-    public double BEEP_EN;
-    public double GPS_RQ;
+    public int BEEP_EN;
+    public int GPS_RQ;
     public double NO_WIFI;
     public double N_END_HOUR;
     public double N_SLEEP_TIME;
     public double N_START_HOUR;
-    public double PORT_RQ;
+    public int PORT_RQ;
     public int SLEEP_TIME;
     public String WIFI_PSSWD_DEFAULT;
     public String WIFI_SSID_DEFAULT;
     public double aux;
 
     public void confvalues(){}
-    public void confvalues(double BAT_H, double BAT_L, double BEEP_EN, double GPS_RQ, double NO_WIFI, double N_END_HOUR, double N_SLEEP_TIME, double N_START_HOUR, double PORT_RQ, int SLEEP_TIME, String WIFI_PSSWD_DEFAULT, String WIFI_SSID_DEFAULT,double aux){
+    public void confvalues(double BAT_H, double BAT_L, int BEEP_EN, int GPS_RQ, double NO_WIFI, double N_END_HOUR, double N_SLEEP_TIME, double N_START_HOUR, int PORT_RQ, int SLEEP_TIME, String WIFI_PSSWD_DEFAULT, String WIFI_SSID_DEFAULT,double aux){
         this.BAT_H = BAT_H;
         this.BAT_L = BAT_L;
         this.BEEP_EN = BEEP_EN;
